extract `uniqBy` as generic method and reuse internally
