+++ packages/ember-runtime/lib/utils.js
@@ -1,8 +1,9 @@
 import { DEBUG } from '@glimmer/env';
-import { PROXY_CONTENT } from 'ember-metal';
+import { PROXY_CONTENT, get } from 'ember-metal';
 import { HAS_NATIVE_PROXY } from 'ember-utils';
-import EmberArray from './mixins/array';
+import EmberArray, { A } from './mixins/array';
 import EmberObject from './system/object';
+import { assert } from '@ember/debug';
 
 // ........................................
 // TYPING & ARRAY MESSAGING
@@ -165,6 +166,26 @@ export function typeOf(item) {
       ret = 'date';
     }
   }
+
+  return ret;
+}
+
+const identityFunction = item => item;
+
+export function uniqBy(array, key = identityFunction) {
+  assert(`first argument passed to \`uniqBy\` should be array`, isArray(array));
+
+  let ret = A();
+  let seen = new Set();
+  let getter = typeof key === 'function' ? key : get;
+
+  array.forEach(item => {
+    let val = getter(item, key);
+    if (!seen.has(val)) {
+      seen.add(val);
+      ret.push(item);
+    }
+  });
 
   return ret;
 }
